requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
easyocr>=1.7.0
Pillow>=10.0.0
python-docx>=1.0.0
pandas>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.0.1
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0